Add script to perform the margin distribution calculation, and its normalization.
This script will compute margin distributions for neural networks. It is
based both on the paper by Jiang et al, and the python library by
Mostafa Elraby, which implements margin distributions in tensorflow.

The paper is available here:
https://arxiv.org/abs/1810.00113

And the library
https://github.com/mostafaelaraby/generalization-gap-features-tensorflow

This research was used as the Basis for the **P**redicting
**G**eneralization in **D**eep **L**earning competition, from Neurips
2020.

Information about this competition is available here:

https://arxiv.org/abs/2012.07976`
https://competitions.codalab.org/competitions/25301